requirements.txt
## Core
python-dotenv
pandas
Levenshtein
SQLAlchemy
psycopg[binary,pool]

# Jupyter
ipykernel
jupyter_black

# Formatters / linters
types-setuptools
pandas-stubs
black
mypy
isort

# App
discord.py